Keep ChangeLock inteface and its only implementation in one file
